+++ lib/secure_session.dart
@@ -22,34 +22,73 @@ final class CookieOptions {
     this.sameSite,
   });
 
+  /// The path of the cookie
   final String path;
 
+  /// The domain of the cookie
   final String? domain;
 
+  /// The expiry of the cookie
   final DateTime? expires;
 
+  /// The max age of the cookie
   final int? maxAge;
 
+  /// Whether the cookie is secure
   final bool secure;
 
+  /// Whether the cookie is http only
   final bool httpOnly;
 
+  /// The SameSite attribute of the cookie
   final SameSite? sameSite;
 }
 
-/// A class to manage secure sessions
-class SecureSession {
-  /// Creates a new instance of [SecureSession]
-  SecureSession({
-    String? cookieName,
+/// A class to encapsulate the options for a session
+final class SessionOptions {
+  /// The name of the session
+  final String defaultSessionName;
+
+  /// The name of the cookie
+  final Duration expiry;
+
+  /// The separator for the session
+  final String separator;
+
+  /// The options for the cookie
+  final CookieOptions cookieOptions;
+
+  /// The secret for the session
+  final String? secret;
+
+  /// The salt for the session
+  final String? salt;
+
+  /// The name of the cookie
+  final String? cookieName;
+
+  /// The path to the key
+  final String? keyPath;
+
+  late String _key;
+
+  /// The key for the session
+  String get key => _key;
+
+  /// Creates a new instance of [SessionOptions]
+  SessionOptions({
+    this.cookieName,
     this.defaultSessionName = 'session',
     this.expiry = const Duration(days: 1),
-    String? keyPath,
+    this.keyPath,
     this.separator = r';',
     this.secret,
     this.cookieOptions = const CookieOptions(),
     this.salt,
-  }) : cookieName = cookieName ?? defaultSessionName {
+  }) {
+    if (secret == null && keyPath == null) {
+      throw ArgumentError('Either secret or keyPath must be provided');
+    }
     if (keyPath == null && secret == null) {
       throw ArgumentError('Either key or secret must be provided');
     }
@@ -57,7 +96,7 @@ class SecureSession {
       throw ArgumentError('Salt is only used with secret');
     }
     if (keyPath != null) {
-      final filePath = File(keyPath);
+      final filePath = File(keyPath!);
       if (!filePath.existsSync()) {
         throw ArgumentError('Key file does not exist');
       }
@@ -81,36 +120,19 @@ class SecureSession {
       _key = secret!;
     }
   }
+}
 
-  /// The name of the cookie
-  ///
-  /// Default is the value of [defaultSessionName]
-  final String? cookieName;
-
-  /// The name of the default session
-  final String defaultSessionName;
-
-  /// The separator used to separate the cipher and nonce
-  final String separator;
-
-  /// The expiry of the session
-  final Duration expiry;
-
-  /// The options for the cookie
-  final CookieOptions cookieOptions;
-
-  String? _key;
-
-  /// The key used to encrypt the session
-  String? get key {
-    return _key;
+/// A class to manage secure sessions
+class SecureSession {
+  /// Creates a new instance of [SecureSession]
+  SecureSession({required this.options}) {
+    if (options.isEmpty) {
+      throw ArgumentError('At least one session option must be provided');
+    }
   }
 
-  /// The secret used to encrypt the session
-  final String? secret;
-
-  /// The salt used to encrypt the session
-  final String? salt;
+  /// The options for the session
+  final List<SessionOptions> options;
 
   final Map<String, SessionValue> _data = {};
 
@@ -120,41 +142,30 @@ class SecureSession {
   /// [session] is the current session
   ///
   /// Usable inside your request handler
-  void init(List<Cookie> cookies, HttpSession session) {
-    final sessionCookie =
-        cookies.where((cookie) => cookie.name == cookieName).firstOrNull;
-    final sessionValues = cookies.where(
-        (cookie) => session.keys.any((session) => session == cookie.name));
-    sessionCookie?.domain ??= cookieOptions.domain;
-    sessionCookie?.expires ??= cookieOptions.expires;
-    sessionCookie?.httpOnly = cookieOptions.httpOnly;
-    sessionCookie?.maxAge ??= cookieOptions.maxAge;
-    sessionCookie?.path ??= cookieOptions.path;
-    sessionCookie?.secure = cookieOptions.secure;
-    sessionCookie?.sameSite ??= cookieOptions.sameSite;
-    if (sessionCookie != null) {
-      final sessionData = sessionCookie.value;
-      final sessionValue = decode(sessionData);
-      if (sessionValue != null && sessionValue.isNotEmpty) {
-        final split = sessionValue.split(separator);
-        final value = split[0];
-        final ttl =
-            DateTime.fromMillisecondsSinceEpoch(int.tryParse(split[1]) ?? 0)
-                .difference(DateTime.now())
-                .inMilliseconds;
-        _data[defaultSessionName] = SessionValue(value, ttl);
-      }
-    }
-    for (final cookie in sessionValues) {
-      final sessionValue = decode(cookie.value);
-      if (sessionValue != null && sessionValue.isNotEmpty) {
-        final split = sessionValue.split(separator);
-        final value = split[0];
-        final ttl =
-            DateTime.fromMillisecondsSinceEpoch(int.tryParse(split[1]) ?? 0)
-                .difference(DateTime.now())
-                .inMilliseconds;
-        _data[cookie.name] = SessionValue(value, ttl);
+  void init(List<Cookie> cookies) {
+    for (final option in options) {
+      final cookieName = option.cookieName ?? option.defaultSessionName;
+      final sessionCookie =
+          cookies.where((cookie) => cookie.name == cookieName).firstOrNull;
+      sessionCookie?.domain ??= option.cookieOptions.domain;
+      sessionCookie?.expires ??= option.cookieOptions.expires;
+      sessionCookie?.httpOnly = option.cookieOptions.httpOnly;
+      sessionCookie?.maxAge ??= option.cookieOptions.maxAge;
+      sessionCookie?.path ??= option.cookieOptions.path;
+      sessionCookie?.secure = option.cookieOptions.secure;
+      sessionCookie?.sameSite ??= option.cookieOptions.sameSite;
+      if (sessionCookie != null) {
+        final sessionData = sessionCookie.value;
+        final sessionValue = decode(sessionData, option, false);
+        if (sessionValue != null && sessionValue.isNotEmpty) {
+          final split = sessionValue.split(option.separator);
+          final value = split[0];
+          final ttl =
+              DateTime.fromMillisecondsSinceEpoch(int.tryParse(split[1]) ?? 0)
+                  .difference(DateTime.now())
+                  .inMilliseconds;
+          _data[cookieName] = SessionValue(value, ttl, option);
+        }
       }
     }
   }
@@ -163,36 +174,88 @@ class SecureSession {
   ///
   /// [value] can be a string or a json serializable object
   /// [sessionName] is the name of the session to write to (default is the value of [defaultSessionName])
-  void write(dynamic value, {String? sessionName}) {
-    final name = sessionName ?? defaultSessionName;
+  void write(dynamic value, String sessionName, [SessionOptions? opts]) {
+    final option = options
+            .where((e) => (e.cookieName ?? e.defaultSessionName) == sessionName)
+            .firstOrNull ??
+        opts;
+    if (option == null) {
+      throw ArgumentError('Session not found');
+    }
+    final name = (option.cookieName ?? option.defaultSessionName);
     if (value is! String &&
         value is! List<Map<String, dynamic>> &&
         value is! Map<String, dynamic>) {
       throw ArgumentError(
           'Value must be a string or a json serializable object');
     }
-    _data[name] = encode(value);
+    if (_data.containsKey(name)) {
+      _data[name]!.deleted = false;
+    }
+    _data[name] = encode(value, option);
   }
 
   /// Reads a value from the session
   ///
   /// [sessionName] is the name of the session to read from (default is the value of [defaultSessionName])
   /// Returns the value of the session or null if the session does not exist or has expired
-  String? read([String? sessionName]) {
-    final name = sessionName ?? defaultSessionName;
+  String? read(String sessionName, [SessionOptions? opts]) {
+    final option = options
+            .where((e) => (e.cookieName ?? e.defaultSessionName) == sessionName)
+            .firstOrNull ??
+        opts;
+    if (option == null) {
+      throw ArgumentError('Session not found');
+    }
+    final name = (option.cookieName ?? option.defaultSessionName);
     if (!_data.containsKey(name)) {
       return null;
     }
-    return decode(_data[name]!.value);
+    if (_data[name]!.deleted) {
+      return null;
+    }
+    return decode(_data[name]!.value, option, _data[name]!.hasChanged);
+  }
+
+  /// Deletes a session
+  void delete(String sessionName, [SessionOptions? opts]) {
+    final option = options
+            .where((e) => (e.cookieName ?? e.defaultSessionName) == sessionName)
+            .firstOrNull ??
+        opts;
+    if (option == null) {
+      throw ArgumentError('Session not found');
+    }
+    final name = (option.cookieName ?? option.defaultSessionName);
+    if (!_data.containsKey(name)) {
+      return;
+    }
+    _data[name]!.deleted = true;
+  }
+
+  /// Regenerates the session
+  void regenerate(String sessionName, [SessionOptions? opts]) {
+    final option = options
+            .where((e) => (e.cookieName ?? e.defaultSessionName) == sessionName)
+            .firstOrNull ??
+        opts;
+    if (option == null) {
+      throw ArgumentError('Session not found');
+    }
+    final name = (option.cookieName ?? option.defaultSessionName);
+    if (!_data.containsKey(name)) {
+      return;
+    }
+    _data[name]!.hasChanged = true;
   }
 
   /// Utility function to decode a value
   ///
   /// [value] is a string to decode
   /// Returns a [String] object
-  String? decode(String value) {
+  String? decode(String value, SessionOptions options, bool hasChanged) {
     /// Split the value into cipher and nonce
-    final splittedValue = value.split(separator);
+    final splittedValue = value.split(options.separator);
 
     /// If the value does not contain a cipher and nonce then return an empty string
     if (splittedValue.length != 2) {
@@ -206,14 +269,14 @@ class SecureSession {
     if (cipher.isEmpty || nonce.length < 16) {
       return null;
     }
-    final encrypter = Fernet(Key.fromUtf8(_key! + utf8.decode(nonce)));
+    final encrypter = Fernet(Key.fromUtf8(options.key + utf8.decode(nonce)));
 
     /// Decrypt the cipher
     final decrypted = encrypter.decrypt(Encrypted.from64(cipher),
-        ttl: DateTime.now().add(expiry).millisecondsSinceEpoch);
+        ttl: DateTime.now().add(options.expiry).millisecondsSinceEpoch);
 
     /// Split the decrypted value into payload and timestamp
-    final separatedPayload = utf8.decode(decrypted).split(separator);
+    final separatedPayload = utf8.decode(decrypted).split(options.separator);
     if (separatedPayload.length != 2) {
       return null;
     }
@@ -224,7 +287,7 @@ class SecureSession {
         .inMilliseconds;
 
     /// If the timestamp is greater than the expiry then return null
-    if (ts > expiry.inMilliseconds) {
+    if (ts > options.expiry.inMilliseconds && !hasChanged) {
       return null;
     }
     return payload;
@@ -237,17 +300,17 @@ class SecureSession {
   /// Throws an [ArgumentError] if the value already contains the separator
   ///
   /// The value is encrypted using the Fernet algorithm
-  SessionValue encode(dynamic value) {
+  SessionValue encode(dynamic value, SessionOptions options) {
     final msg = value is String ? value : jsonEncode(value);
-    if (msg.contains(separator)) {
+    if (msg.contains(options.separator)) {
       throw ArgumentError('Value cannot contain the separator');
     }
-    final nonce = salt ?? _generateNonce();
-    final encrypter = Fernet(Key.fromUtf8(_key! + nonce));
+    final nonce = options.salt ?? _generateNonce();
+    final encrypter = Fernet(Key.fromUtf8(options.key + nonce));
     final ts = DateTime.now().millisecondsSinceEpoch;
     final cipher = encrypter.encrypt(utf8.encode('$msg\$$ts'));
     return SessionValue(
-        '${cipher.base64};${base64Url.encode(nonce.codeUnits)}', ts);
+        '${cipher.base64};${base64Url.encode(nonce.codeUnits)}', ts, options);
   }
 
   String _generateNonce() {
@@ -263,14 +326,27 @@ class SecureSession {
   void clear() {
     _data.clear();
   }
+
+  Map<String, SessionValue> get data => _data;
 }
 
+/// A class to encapsulate a session value
 class SessionValue {
-  SessionValue(this.value, this.ttl);
+  /// Creates a new instance of [SessionValue]
+  SessionValue(this.value, this.ttl, this.options);
 
+  /// The options for the session
+  SessionOptions options;
+
+  /// Whether the session has been deleted
+  bool deleted = false;
+
+  /// The value of the session
   dynamic value;
 
+  /// Whether the session has changed
   bool hasChanged = false;
 
+  /// The time to live of the session
   int ttl;
 }