+++ CHANGELOG.md
@@ -7,3 +7,9 @@
 ## 1.0.1
 
 - Add the `clear` method.
+
+## 1.0.2
+
+- Add the `delete` method.
+- Add the `regenerate` method.
+- Add the possibility to set multiple session options.